% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_VCF_functions.R
\name{add_DBS_cats_ToVCF}
\alias{add_DBS_cats_ToVCF}
\title{add_DBS_cats_ToVCF}
\usage{
add_DBS_cats_ToVCF(vcf = NULL, DBS_mutations_only = NA)
}
\arguments{
\item{DBS_mutations_only}{Logical, TRUE if all lines in input VCF correspond to double base substitutions, FALSE if it contains other SNVs and/or Indels.}

\item{input}{Input VCF to which DBS mutation categories are to be added.}
}
\description{
Adds DBS mutation categories to a VCF containing SNVs. N.B. The VCF must be ordered by sample, CHROM and position for this function to work.
}
\examples{
vcf <- add_DBS_cats_ToVCF(input = vcf, DBS_mutations_only = FALSE)
}
\keyword{Signatures}
