% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_VCF_functions.R
\name{add_ID_cats_ToVCF}
\alias{add_ID_cats_ToVCF}
\title{add_ID_cats_ToVCF}
\usage{
add_ID_cats_ToVCF(vcf = NULL, ref_fasta = NULL, palimpdir_man = NA)
}
\arguments{
\item{vcf}{VCF to which Indel mutation categories are to be added.}

\item{ref_fasta}{Path to fasta file for reference genome of choice (e.g. hg19 genome).}

\item{palimpdir_man}{Path to palimpsest directory entered manually.}

\item{tool_dir}{Path to folder containing PCAWG7-data-preparation-version-1.5 tool.}
}
\description{
Adds Indel mutation categories to a VCF.
}
\examples{
vcf <- add_ID_cats_ToVCF(vcf = vcf,ref_fasta = ref_fasta)
}
\keyword{Signatures}
