% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_timings.R
\name{chrTime_annot}
\alias{chrTime_annot}
\title{chrTime_annot}
\usage{
chrTime_annot(vcf = NULL, cna_data = NULL, cyto = cyto)
}
\arguments{
\item{vcf}{vcf data frame with mutation information to annotate}

\item{cna_data}{Data frame with Copy Number Alteration Information}

\item{cyto}{Cytoband description table}
}
\description{
Function to annotate the mutation data with chromosomal duplication timing information
}
