% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_plotting.R
\name{chrTime_plot}
\alias{chrTime_plot}
\title{chrTime_plot}
\usage{
chrTime_plot(vcf = NULL, point.mut.time = NULL, resdir = resdir)
}
\arguments{
\item{vcf}{vcf data frame containing the mutations with multiplicity annotation, as obtained using chrTime_annot function}

\item{point.mut.time}{Timing of chromosome duplications in point mutation time, as obtained using chrTime_annot function}

\item{resdir}{Result directory}
}
\description{
Function to represent the timing of chromosomal duplications in point mutation time
}
