% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_utils.R
\name{cit.chromString2num}
\alias{cit.chromString2num}
\title{cit.chromString2num}
\usage{
cit.chromString2num(d, chrom = "chrom", chromNum = "chrNum")
}
\arguments{
\item{d}{data.frame which columns include genomic position information}

\item{chrom}{the column in \code{d} giving the chromosome position}
}
\description{
convert chromosomes X, Y and MT into numerics 23, 24 and 25
}
