% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_utils.R
\name{cit.density}
\alias{cit.density}
\title{cit.density}
\usage{
cit.density(x, doplot = FALSE, pc = 0.05, ...)
}
\arguments{
\item{x}{a numeric vector}

\item{doplot}{boolean}

\item{...}{}
}
\description{
wrapping of the function 'density' adding down and top points to the result
}
