% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_utils.R
\name{cit.discretize}
\alias{cit.discretize}
\title{cit.discretize}
\usage{
cit.discretize(x, lim, quant = FALSE, addlevels = FALSE)
}
\arguments{
\item{x}{a vector}

\item{lim}{cut-off(s)}

\item{quant}{if TRUE (default FALSE) the \code{x} is discretize by quantile and \code{lim} is considered as cut-off(s) for quantile, ie 0<lim<1}

\item{addlevels}{add character levels indicating the cut-offs (i.e. for un cut-off iqq levels=c("<iqq",">=iqq") )}
}
\description{
Discretize a continous variable by specified \code{lim} cut-off(s)
}
