% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_utils.R
\name{cit.genomOrder}
\alias{cit.genomOrder}
\title{cit.genomOrder}
\usage{
cit.genomOrder(d, chrom = "chrom", pos = "pos", startPos = "start",
  endPos = "end", chromNum = "chrNum")
}
\arguments{
\item{d}{data.frame}

\item{chrom}{chromosome column name}

\item{pos}{sequence location column name}

\item{startPos}{the column in d giving the start base pair position}

\item{endPos}{the column in d giving the end base pair position}

\item{chromNum}{the name of added column whith chromosome from 1 to 24}
}
\description{
cit.genomOrder
}
