% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_utils.R
\name{cit.pangenomCoord}
\alias{cit.pangenomCoord}
\title{cit.pangenomCoord}
\usage{
cit.pangenomCoord(x, chrom = "chrom", pos = "meanPos", startPos = "start",
  endPos = "end", cyto = NULL, absPos = "absPos", chromNum = "chrNum")
}
\arguments{
\item{x}{data.frame which columns include genomic position information}

\item{chrom}{the column in x giving the chromosome position (in 1:22,X,Y)}

\item{pos}{the column in x giving the base pair position   (optional)}

\item{startPos}{the column in x giving the start base pair position}

\item{endPos}{the column in x giving the end base pair position}

\item{cyto}{cytoband object (optional)}

\item{absPos}{name of added column giving absolute pangenomic order}

\item{chromNum}{the column in x giving the chromosome position with X = 23 & Y =24}
}
\description{
cit.pangenomCoord
}
