% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_utils.R
\name{cit.peaks}
\alias{cit.peaks}
\title{cit.peaks}
\usage{
cit.peaks(
  x,
  percentHighestPeak = 0.2,
  maxNbPeaks = NULL,
  minDeltaBetweenPeaks = 0.03,
  deltaApproach = 1,
  doplot = FALSE,
  ...
)
}
\arguments{
\item{x}{a numeric vector}

\item{percentHighestPeak}{percentHighestPeak}

\item{maxNbPeaks}{maxNbPeaks}

\item{minDeltaBetweenPeaks}{minDeltaBetweenPeaks}

\item{deltaApproach}{deltaApproach}

\item{doplot}{boolean}

\item{...}{}
}
\description{
wrapping of the function 'cit.density' to control the maximum number of peaks
}
