% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_clonality.R
\name{cnaCCF_annot}
\alias{cnaCCF_annot}
\title{cnaCCF_annot}
\usage{
cnaCCF_annot(
  vcf = NULL,
  annot_data = NULL,
  cna_data = NULL,
  CCF_boundary = 0.95
)
}
\arguments{
\item{vcf}{vcf data frame containing somatic mutations}

\item{annot_data}{Annotation file with Gender information and Tumor Purity for each sample}

\item{cna_data}{Data frame with Copy Number Alteration Information}

\item{CCF_boundary}{percentage boundary threshold for clonal mutations}
}
\description{
This function allows to annotate the input mutation data with clonality information (Cancer Cell Fraction- CCF) based on the Copy Number information provided
}
