% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_plotting.R
\name{cnaCCF_plots}
\alias{cnaCCF_plots}
\title{cnaCCF_plots}
\usage{
cnaCCF_plots(vcf = NULL, resdir = resdir)
}
\arguments{
\item{vcf}{vcf data frame containing the mutations}

\item{resdir}{Result directory}
}
\description{
cnaCCF_plots
}
