% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COSMIC_Signatures}
\alias{COSMIC_Signatures}
\title{COSMIC_Signatures}
\format{A data frame with 30 observations and 96 variables:
\describe{
  \item{signature}{COSMIC signatures}
  \item{mutation type}{mutation belonging to 96 mutational context}
  ...
}}
\source{
\url{Palimpsest}
}
\usage{
COSMIC_Signatures
}
\description{
A dataset containing the proportions of mutations for already described mutational
signatures as described in COSMIC database
}
\keyword{datasets}
