% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_input2_function.R
\name{count_triplets}
\alias{count_triplets}
\title{count_triplets}
\usage{
count_triplets(vcf, ordering, mutypes)
}
\arguments{
\item{vcf}{The input vcf file.}

\item{ordering}{the ordering of samples from palimpsest_input2 function.}

\item{mutypes}{The DBS mutypes from palimpsest_input2 function.}
}
\description{
Counts the number of DBS mutations that are in fact triple base substituions. Used inside "palimpsest_input2", not designed to be used outside this function.
}
\examples{
trip_nums <- count_triplets(vcf, ordering, mutypes)
}
\keyword{Signatures}
