% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_deconvolution.R
\name{deconvolution_compare}
\alias{deconvolution_compare}
\title{deconvolution_compare}
\usage{
deconvolution_compare(new_signatures, COSMIC_Signatures)
}
\arguments{
\item{new_signatures}{Data frame of de-novo extracted mutational signatures}

\item{COSMIC_Signatures}{Data frame of already published mutational signatures (Example: Alexandrov et.al,2013)}
}
\description{
Function to calculate cosine similarity scores between two sets of mutational signatures.
}
