% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_compare2_function.R
\name{deconvolution_compare2}
\alias{deconvolution_compare2}
\title{deconvolution_compare2}
\usage{
deconvolution_compare2(new_signatures, COSMIC_Signatures)
}
\arguments{
\item{new_signatures}{Data frame of de novo extracted mutational signatures}

\item{COSMIC_Signatures}{Data frame of reference mutational signatures (The reference SBS_cosmic, DBS_cosmic & ID_cosmic matrices are taken from Alexandrov et al. (2018)).}
}
\description{
Function to calculate cosine similarity scores between two sets of mutational signatures (e.g. SBS COSMIC signatures vs Palimpsest de novo signatures).
}
\examples{
SBS_cosine_similarities <- deconvolution_compare2(SBS_denovo_sigs,SBS_cosmic)
}
