% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_plotting.R
\name{deconvolution_exposure}
\alias{deconvolution_exposure}
\title{deconvolution_exposure}
\usage{
deconvolution_exposure(mutSign_nums, mutSign_props, sig_cols = colclust)
}
\arguments{
\item{mutSign_nums}{Matrix in sample x mutational signature exposure format in numbers}

\item{mutSign_props}{Matrix in sample x mutational signature exposure format in proportions}

\item{sig_cols}{Character vector indicating the colors representing each signature in graphical outputs. Must match to the total number of provided signatures}
}
\description{
Function to plot the exposure of mutational signatures in samples across the entire series
}
