% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_fit2_function.R
\name{deconvolution_fit2}
\alias{deconvolution_fit2}
\title{deconvolution_fit2}
\usage{
deconvolution_fit2(input_matrices = NULL, input_signatures = NULL,
  threshold = 6, sig_cols = NA, doplot = TRUE,
  save_signatures_exp = TRUE, resdir = resdir)
}
\arguments{
\item{input_matrices}{Palimpsest input list of mutation number and proportion matrices.}

\item{input_signatures}{Matrix of the mutational signatures to fit within the provided cohort of samples.}

\item{threshold}{Signatures contributing less then this percentage of total mutations within a sample will be discarded (e.g. if set to 6 and signature X contributes 5% of a sample's mutations, signature X will not be reported as being present in this sample).}

\item{sig_cols}{Character vector of R-compatable colours representing each signature to be used graphical outputs. Each signature in input_signatures must have named colour in this vector for grpahical outputs to work.}

\item{doplot}{Logical indicating whether graphical outputs should be generated (defaults to TRUE).}

\item{save_signatures_exp}{Logical indicating whether or not signatures_exp object should be saved in the redsir (defaults to TRUE).}

\item{resdir}{Results directory.}
}
\description{
Function to calculate the number and proportion of each signature in each sample, in addition to plotting each sample's mutational profile and its signature contribution.
}
\examples{
signatures_exp <- deconvolution_fit2(input_matrices = SBS_input,
                                         threshold = 6,input_signatures = SBS_liver,
                                           sig_cols = sig_cols,plot = T,resdir = resdir)

}
\keyword{Signatures}
