% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_deconvolution.R
\name{deconvolution_nmf}
\alias{deconvolution_nmf}
\title{deconvolution_nmf}
\usage{
deconvolution_nmf(input_data = NULL, type = NULL, range_of_sigs = NULL,
  num_of_sigs = "auto", nrun = nrun, method = "brunet",
  plot_sigs = TRUE, resdir = resdir)
}
\arguments{
\item{input_data}{Matrix in mutation type x sample format in proportions}

\item{type}{SNV for single nucleotide variant signatures, SV for structural variant signatures}

\item{range_of_sigs}{numeric Range of possible numbers of mutational signatures to extract}

\item{num_of_sigs}{Number of mutational signatures to extract. If left to default "auto" value, the appropriate number of signatures will be estimated from NMF metrics.}

\item{nrun}{Number of iterations to perform for each number of signatures}

\item{method}{Specification of the NMF algorithm. ‘brunet’ method corresponds to the standard NMF algorithm from Brunet,2004}

\item{plot_sigs}{If TRUE, will output the extracted mutational signature plots}

\item{resdir}{Result directory}
}
\description{
Function to perform de novo mutational signature extraction and represent the extracted signatures.
}
