% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ensgene}
\alias{ensgene}
\title{ensgene}
\format{An object of class \code{data.frame} with 57849 rows and 11 columns.}
\source{
\url{Palimpsest}
}
\usage{
ensgene
}
\description{
Ensembl Gene Information
}
\keyword{datasets}
