% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_VCF_function.R
\name{extract_dbs_from_vcf}
\alias{extract_dbs_from_vcf}
\title{extract_dbs_from_vcf}
\usage{
extract_dbs_from_vcf(vcf = NULL)
}
\arguments{
\item{vcf}{The input VCF from which DBS mutations are to be extracted}
}
\description{
returns lines of a VCF corresponding to DBS mutations. The VCF must be ordered by sample, CHROM and position for this function to work (can be performed by the "order_vcf()" function.
}
\examples{
vcf_dbs <- extract_dbs_from_vcf(vcf = vcf)
}
\keyword{Signatures}
