% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_utils.R
\name{factoall}
\alias{factoall}
\title{factoall}
\usage{
factoall(d, ncmax = 10)
}
\arguments{
\item{d}{Data frame to be converted}

\item{ncmax}{Maximum number of characters for numeric fields}
}
\description{
Function to convert factors in a data frame to the appropriate format (character, numeric...)
}
