% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_utils.R
\name{factochar}
\alias{factochar}
\title{factochar}
\usage{
factochar(d)
}
\arguments{
\item{d}{Data frame to be converted}
}
\description{
Function to convert factors factors in a data frame to characters
}
