% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_utils.R
\name{\%!in\%}
\alias{\%!in\%}
\title{"Not in" function}
\usage{
x \%!in\% y
}
\description{
The opposite of << %in% >>, used by several palimpsest functions.
}
\examples{
days <- c("monday","tuesday","wednesday","thursday","friday","saturday","sunday")
weekdays <- days[days \%!in\% c("saturday", "sunday")]
}
\keyword{Signatures}
