% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indel_category_correct_function.R
\name{indel_category_correct}
\alias{indel_category_correct}
\title{indel_category_correct}
\usage{
indel_category_correct(input = NULL)
}
\arguments{
\item{input}{VCF of indel mutations}
}
\description{
Corrects Indel category output from PCAWG7-data-preparation-version-1.5
}
\examples{
vcf_ID <- indel_category_correct(input = vcf)
}
\keyword{Signatures}
