% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_plotting.R
\name{make_plot}
\alias{make_plot}
\title{make_plot}
\usage{
make_plot(translocs, others, Sample_to_plot, resdir)
}
\arguments{
\item{translocs}{translocs}

\item{others}{others}

\item{Sample_to_plot}{Sample_to_plot}

\item{resdir}{resdir}
}
\description{
make_plot
}
