% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_extraction.R
\name{NMF_Extraction}
\alias{NMF_Extraction}
\title{NMF_Extraction}
\usage{
NMF_Extraction(input_matrices = NULL, range_of_sigs = NULL,
  num_of_sigs = "auto", nrun = 10, method = "brunet",
  plot_sigs = TRUE, resdir = NA)
}
\arguments{
\item{input_matrices}{Palimpsest input list of mutation number and proportion matrices.}

\item{range_of_sigs}{Numerical range of signatures. If "num_of_sigs" is set to auto, NMF will estiamte the optimal number of signatures in the input within this range before extracting with this number.}

\item{num_of_sigs}{The number of mutational signatures to extract. If left to default "auto" value, the appropriate number of signatures will be estimated from NMF metrics (which are plotted in resdir). "auto" is recommended on the first run, but once the optimal number of signatures is known, setting this argument to that value will make the extraction quicker.}

\item{method}{Specification of the NMF algorithm. ‘brunet’ method corresponds to the standard NMF algorithm from Brunet et al. (2004, PNAS).}

\item{plot_sigs}{Logical, if TRUE (default), the extracted mutational signatures will be plotted in the resdir.}

\item{resdir}{Results directory.}
}
\description{
Extracts mutational signatures de novo using NMF. Also estimates the optimal number of mutational signatures in the input.
}
\examples{
SBS_denovo_signatures <- NMF_Extraction(input_matrices = SBS_input, range_of_sigs = 1:20,nrun = 10,resdir = resdir)
}
\keyword{Signatures}
