% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_vcf_function.R
\name{order_vcf}
\alias{order_vcf}
\title{order_vcf}
\usage{
order_vcf(vcf, Project_col = NA)
}
\arguments{
\item{vcf}{The VCF to be ordered.}

\item{Project_col}{Name of the Project column in the VCF (e.g. may contain "ICGC")}
}
\description{
Orders a VCF by position, chromosome, sample and project (if project argument given).
}
\examples{
vcf <- order_vcf(vcf, Project_col = "Project")
}
\keyword{Signatures}
