% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_utils.R
\name{palimpsest_addMutationContextToVcf}
\alias{palimpsest_addMutationContextToVcf}
\title{palimpsest_addMutationContextToVcf}
\usage{
palimpsest_addMutationContextToVcf(vcf, Reference_Genome, chrom.col = "chr",
  start.col = "start", end.col = "end", ref.col = "ref",
  alt.col = "alt", strand.mut.col = "strand.mut",
  strand.gene.col = "strand.gene", sample.col = "samplenames",
  strand.ts.output.col = "strand.ts", substype.output.col = "substype",
  ctx3.output.col = "context3", ctx5.output.col = "context5",
  mutcat3.output.col = "mutcat3", mutcat5.output.col = "mutcat5")
}
\arguments{
\item{vcf}{vcf data frame containing the SNVs}

\item{Reference_Genome}{Reference genome (e.g. BSgenome.Hsapiens.UCSC.hg19)}

\item{chrom.col}{Name of column containing chromosomes (should be in chr1, chr2... format)}

\item{start.col}{Name of column containing start position}

\item{end.col}{Name of column containing end position}

\item{ref.col}{Name of column containing reference base}

\item{alt.col}{Name of column containing mutated base}

\item{strand.mut.col}{Name of column containing mutation strand}

\item{strand.gene.col}{Name of column containing gene strand}

\item{sample.col}{Name of column containing sample names}

\item{strand.ts.output.col}{Name of added column with transcribed or non transcribed strand}

\item{substype.output.col}{Name of added column with substitution type}

\item{ctx3.output.col}{Name of added column with 3 nucleotide context}

\item{ctx5.output.col}{Name of added column with 5 nucleotide context}

\item{mutcat3.output.col}{Name of added column with 3 nucleotide categories (includes substitution type and context)}

\item{mutcat5.output.col}{Name of added column with 5 nucleotide categories (includes substitution type and context)}
}
\description{
Function to add mutation contexts to vcf
}
