% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_VCF_function.R
\name{palimpsest_addMutationContextToVR_2}
\alias{palimpsest_addMutationContextToVR_2}
\title{palimpsest_addMutationContextToVR_2}
\usage{
palimpsest_addMutationContextToVR_2(vr = NULL, ref_genome = NULL,
  k = 3, check.strand = FALSE, num_of_DELs = NULL)
}
\value{
vcf
}
\description{
Function used within annotate_VCF to add mutation context to SBS mutations.
}
\examples{
vr3 <- palimpsest_addMutationContextToVR_2(vr = vr, ref_genome = ref_genome, k = 3, unify = TRUE,num_of_DELs = delet_this)
}
