% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_functions.R
\name{palimpsest_addSVcategoriesToVcf}
\alias{palimpsest_addSVcategoriesToVcf}
\title{palimpsest_addSVcategoriesToVcf}
\usage{
palimpsest_addSVcategoriesToVcf(sv = sv, type.col = "Type",
  sample.col = "Sample", CHROM_1.col = "CHROM",
  CHROM_2.col = "CHROM", POS_1.col = "POS", POS_2.col = "POS",
  resdir = resdir, genome_build = "hg19")
}
\arguments{
\item{sv}{vcf data frame containing the SVs}

\item{type.col}{SV type description column name in mutation_data ["DEL","DUP","INV"]}

\item{sample.col}{Sample column name in sv}

\item{CHROM_1.col}{Start chromosome column name in sv}

\item{CHROM_2.col}{End chromosome column name in sv}

\item{POS_1.col}{Start position column name in sv}

\item{POS_2.col}{End position column name in sv}

\item{resdir}{Result directory}
}
\description{
Function to add SV mutation categories to vcf
}
