% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_plotting.R
\name{palimpsest_CCFplot}
\alias{palimpsest_CCFplot}
\title{palimpsest_CCFplot}
\usage{
palimpsest_CCFplot(vcf = NULL, CHROM.col = "CHROM",
  VAF.col = "Tumor_Varprop", clonality.col = "CCF_Clonality")
}
\arguments{
\item{vcf}{vcf data frame containing the mutations}

\item{CHROM.col}{Chromosome column name in vcf.}

\item{VAF.col}{Variant Allele Fraction information column name in vcf.}

\item{clonality.col}{Clonaility information column name in vcf.}
}
\description{
Function to plot Variant Allele Fraction and CCF distribution in a Sample
}
