% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_plotting.R
\name{palimpsest_clonalitySigsCompare}
\alias{palimpsest_clonalitySigsCompare}
\title{palimpsest_clonalitySigsCompare}
\usage{
palimpsest_clonalitySigsCompare(clonsig = clonsig, subsig = subsig,
  msigcol = msigcol, resdir = resdir)
}
\arguments{
\item{clonsig}{Data frame of sample x clonal mutational signature exposure format in numbers}

\item{subsig}{Data frame of sample x subclonal mutational signature exposure format in numbers}

\item{msigcol}{list of colors to for plotting signature contribution. Must match to the total number of provided signatures}

\item{resdir}{Result directory}
}
\description{
Function to plot comparison between mutational signatures exposures within clonal vs. subclonal mutations in the series.
}
