% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_plotting.R
\name{palimpsest_cnvProfile}
\alias{palimpsest_cnvProfile}
\title{palimpsest_cnvProfile}
\usage{
palimpsest_cnvProfile(vcf = NULL, VAF.color = NULL, cyto = NULL)
}
\arguments{
\item{vcf}{vcf data frame containing the mutations}

\item{VAF.color}{vector of colors for the VAF plot (optionnal).}
}
\description{
Funciton to plot copy number variation plots
}
