% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_plotting.R
\name{palimpsest_DissectSigs}
\alias{palimpsest_DissectSigs}
\title{palimpsest_DissectSigs}
\usage{
palimpsest_DissectSigs(vcf = NULL, signatures_exp_clonal = NULL,
  signatures_exp_subclonal = NULL, sig_cols = NULL, resdir = NULL)
}
\arguments{
\item{vcf}{vcf data frame containing the mutations}

\item{signatures_exp_clonal}{Mutational signature exposures extracted from clonal mutations only using deconvolution_fit function.}

\item{signatures_exp_subclonal}{Mutational signature exposures extracted from subclonal mutations only using deconvolution_fit function.}

\item{sig_cols}{Character vector indicating the colors representing each signature in graphical outputs. Must match to the total number of provided signatures}

\item{resdir}{Result directory}
}
\description{
Function to plot temporal dissection of mutational signatures within a sample
}
