% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_utils.R
\name{palimpsest_distCosine}
\alias{palimpsest_distCosine}
\title{palimpsest_distCosine}
\usage{
palimpsest_distCosine(m)
}
\arguments{
\item{required}{}
}
\description{
Function to calculate cosine distance
}
