% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_plotting.R
\name{palimpsest_Fireworks}
\alias{palimpsest_Fireworks}
\title{palimpsest_Fireworks}
\usage{
palimpsest_Fireworks(
  vcf = NULL,
  CHROM.col = "CHROM",
  POS.col = "POS",
  VAF.col = "Tumor_Varprop",
  LogR.col = "LogR",
  nPurity.col = "nPurity"
)
}
\arguments{
\item{vcf}{vcf data frame containing the mutations}

\item{CHROM.col}{Chromosome column name in vcf.}

\item{POS.col}{Position column name in vcf.}

\item{VAF.col}{Variant Allele Fraction information column name in vcf.}

\item{LogR.col}{LogR information column name in vcf.}

\item{nPurity.col}{Tumor purity column name in vcf.}
}
\description{
Function to plot fireworks plotfor variant allele fractions
}
