% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_input.R
\name{palimpsest_input}
\alias{palimpsest_input}
\title{palimpsest_input}
\usage{
palimpsest_input(vcf = NULL, Type = NULL, mutypes = NA)
}
\arguments{
\item{vcf}{The input VCF file from which a matrix of mutation categories is to be made.}

\item{Type}{Mutation type (SBS, DBS, ID or SV).}

\item{mutypes}{leave blank to use default COSMIC categories (recommeneded), or add your own.}
}
\description{
Creates the list of input matrices of the numbers and proportions of each mutation category (of one mutation type, e.g. SBS mutation categories) in each sample.
}
\examples{
DBS_input <- palimpsest_input(vcf=vcf, Type = "DBS")
}
\keyword{Signatures}
