% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_input2_function.R
\name{Palimpsest_input2}
\alias{Palimpsest_input2}
\title{Palimpsest_input2}
\usage{
Palimpsest_input2(vcf = NULL, Type = NULL, mutypes = NA)
}
\arguments{
\item{vcf}{The input VCF file from which a matrix of mutation categories is to be made.}

\item{Type}{Mutation type (SBS, DBS, ID or SV).}

\item{mutypes}{leave blank to use default COSMIC categories (recommeneded), or add your own.}
}
\description{
Creates the input matrices of the numbers and proportions of each mutation category (of one mutation type, e.g. SBS mutation categories) in each sample.
}
\examples{
DBS_matrix <- Palimpsest_input2(vcf=vcf, Type = "DBS", proportion = TRUE)
}
\keyword{Signatures}
