% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_utils.R
\name{palimpsest_makeMutypeMatFromVcf}
\alias{palimpsest_makeMutypeMatFromVcf}
\title{palimpsest_makeMutypeMatFromVcf}
\usage{
palimpsest_makeMutypeMatFromVcf(vcf, sample.col = "sample",
  mutcat.col = "mutcat3", mutypes = c("CA", "CG", "CT", "TA", "TC", "TG"),
  proportion = TRUE)
}
\arguments{
\item{vcf}{vcf data frame containing the mutations}

\item{sample.col}{Name of column containing sample names}

\item{mutcat.col}{Name of column containing mutation categories}

\item{mutypes}{List of categories to be included in the output}

\item{proportion}{If TRUE, the output matrix will indicate mutation proportions instead of numbers}
}
\description{
Function to create a matrix in mutation type x sample format with either counts or proportions
}
