% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_plotting.R
\name{palimpsest_plotTumorHistories}
\alias{palimpsest_plotTumorHistories}
\title{palimpsest_plotTumorHistories}
\usage{
palimpsest_plotTumorHistories(
  vcf = NULL,
  sv.vcf = NULL,
  cna_data = NULL,
  point.mut.time = NULL,
  clonsig = NULL,
  subsig = NULL,
  msigcol = NULL,
  msigcol.sv = NULL,
  spacechar = 1,
  ystep = 2,
  resdir = NULL
)
}
\arguments{
\item{vcf}{Annotated vcf file for mutations}

\item{sv.vcf}{Annotated vcf file for SVs}

\item{cna_data}{Data frame with Copy Number Alteration Information}

\item{point.mut.time}{Duplication timing estimated from point mutations}

\item{clonsig}{Matrix giving the proportion of clonal mutations attributed to each signature}

\item{subsig}{Matrix giving the proportion of subclonal mutations attributed to each signature}

\item{msigcol}{Color codes for mutational signatures}

\item{msigcol.sv}{Color codes for SV signatures}

\item{spacechar}{Space for characters in the legends}

\item{ystep}{Step between annotation lines}

\item{resdir}{Results directory.}
}
\description{
Function to represent the natural histories of tumors
}
