% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_deconvolution.R
\name{palimpsestOrigin}
\alias{palimpsestOrigin}
\title{palimpsestOrigin}
\usage{
palimpsestOrigin(vcf = vcf, type = "SV", sample.col = "Sample",
  mutcat.col = "Category1",
  signature_contribution = nmf_signatures$sig_nums,
  input_signatures = COSMIC_Signatures)
}
\arguments{
\item{vcf}{data frame containing the mutations/SVs}

\item{type}{SNV for single nucleotide variant signatures, SV for structural variants signatures}

\item{sample.col}{Sample column name in vcf}

\item{mutcat.col}{Mutation category column name in vcf}

\item{signature_contribution}{Matrix in sample x mutational signature exposure format in proportions}

\item{input_signatures}{Data frame describing the mutational signatures to fit within the provided cohort of samples}
}
\description{
Function to estimate the probability of each individual somatic alteration being due to each mutational process
}
