% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_plotting.R
\name{plot.SNV.sigs}
\alias{plot.SNV.sigs}
\title{plot.SNV.sigs}
\usage{
plot.SNV.sigs(spec)
}
\arguments{
\item{spec}{Matrix of mutational signatures x  mutation types}
}
\description{
Function to plot single nucleotide variants signatures (96 nucleotide type)
}
