% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_plotting.R
\name{plot.SV.sigs}
\alias{plot.SV.sigs}
\title{plot.SV.sigs}
\usage{
plot.SV.sigs(sigs)
}
\arguments{
\item{sigs}{Matrix of mutational signatures x  mutation types}
}
\description{
Function to plot structural variants signatures (38 sv categories type)
}
