% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_plotting.R
\name{plot6mutationSpectrumFromVcf}
\alias{plot6mutationSpectrumFromVcf}
\title{plot6mutationSpectrumFromVcf}
\usage{
plot6mutationSpectrumFromVcf(
  vcf,
  sample.col = "sample",
  substype.col = "substype",
  averageProp = FALSE,
  plot.file = NULL
)
}
\arguments{
\item{vcf}{vcf data frame containing the mutations}

\item{sample.col}{Sample column name in vcf}

\item{substype.col}{Substitution types column name in vcf}

\item{averageProp}{If TRUE, the proportions returned will be the average of sample-wise proportions, otherwise the general proportion over the vcf is used (induces slight changes)}

\item{plot.file}{File to which the plot should be saved (if left NULL, plot will be sent directly to the active window)}
}
\description{
Function to plot the proportions of each of the 6 substitution types from a vcf file (containing one or a series of samples)
}
