% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_plotting.R
\name{plot96mutationSpectrumFromVcf}
\alias{plot96mutationSpectrumFromVcf}
\title{plot96mutationSpectrumFromVcf}
\usage{
plot96mutationSpectrumFromVcf(vcf, sample.col = "sample",
  mutcat3.col = "mutcat3", ymax = NULL, averageProp = FALSE,
  plot.file = NULL)
}
\arguments{
\item{vcf}{vcf data frame containing the mutations}

\item{sample.col}{Sample column name in vcf}

\item{mutcat3.col}{Mutation categories column name in vcf}

\item{ymax}{Upper value for the y-axis (if left NULL it will be computed automatically)}

\item{averageProp}{If TRUE, the proportions returned will be the average of sample-wise proportions, otherwise the general proportion over the vcf is used (induces slight changes)}

\item{plot.file}{File to which the plot should be saved (if left NULL, plot will be sent directly to the active window)}
}
\description{
Function to plot the proportions of each of the 96 mutation types from a vcf file (containing one or a series of samples)
}
