% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signatures_function.R
\name{plot_signatures}
\alias{plot_signatures}
\title{plot_signatures}
\usage{
plot_signatures(input_data = NULL, Title = NA, label = "Full")
}
\arguments{
\item{input_data}{Matrix of signature(s) to plot, or an individual sample's proportion of each mutation category. If you receive an input_data related error retry with t(input_data).}

\item{Title}{List (must be same length as number of input signatures) or single title to use. Leave blank to use rownames from input matrix. Use " " for no title.}

\item{label}{"Full" for all labels (default), "Top" for just coloured bars atop the plot; "None" for just bar graph.}
}
\description{
Function to plot signature profiles of all mutation types (SBS, DBS, Indel & SV). 
N.B. recommended width to height ratio of 24:7 must be maintained to keep labels at the correct size.
}
\examples{
pdf(file=paste0(resdir,"SBS_liver_signature_profiles.pdf"),width=24,height=7)
plot_signatures(input_data = SBS_liver,Title = rownames(SBS_liver),Individual = F,label = "Full")
dev.off()
}
\keyword{Signatures}
