% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_plotting.R
\name{plotStrandBias6types}
\alias{plotStrandBias6types}
\title{plotStrandBias6types}
\usage{
plotStrandBias6types(vcf, substype.col = "substype",
  strand.ts.col = "strand.ts", plot.file = NULL)
}
\arguments{
\item{vcf}{vcf data frame containing the mutations}

\item{substype.col}{Substitution types data frame containing the mutations}

\item{strand.ts.col}{Name of column containing indicating whether mutations are on the transcribed or untranscribed strand}

\item{plot.file}{File to which the plot should be saved (if left NULL, plot will be sent directly to the active window)}
}
\description{
Function to calculate and plot the transcriptional strand bias over the 6 substitution types
}
