% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_plotting.R
\name{plotStrandBias96types}
\alias{plotStrandBias96types}
\title{plotStrandBias96types}
\usage{
plotStrandBias96types(
  vcf,
  mutcat3.col = "mutcat3",
  strand.ts.col = "strand.ts",
  plot.file = NULL
)
}
\arguments{
\item{vcf}{data frame containing the mutations}

\item{mutcat3.col}{Mutation categories column name in vcf}

\item{strand.ts.col}{Name of column containing indicating whether mutations are on the transcribed or untranscribed strand}

\item{plot.file}{File to which the plot should be saved (if left NULL, plot will be sent directly to the active window)}
}
\description{
Function to calculate and plot the transcriptional strand bias over the 96 mutation types
}
