% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_plotting.R
\name{plotSubstypeNumberPerSample}
\alias{plotSubstypeNumberPerSample}
\title{plotSubstypeNumberPerSample}
\usage{
plotSubstypeNumberPerSample(
  vcf,
  sample.col = "sample",
  substype.col = "substype",
  plot.file = NULL,
  ...
)
}
\arguments{
\item{vcf}{vcf data frame containing the mutations}

\item{sample.col}{Sample column name in vcf}

\item{substype.col}{Substitution types column name in vcf}

\item{plot.file}{File to which the plot should be saved (if left NULL, plot will be sent directly to the active window)}

\item{...}{}
}
\description{
plotSubstypeNumberPerSample
}
