% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_preprocess.R
\name{preprocessInput_snv}
\alias{preprocessInput_snv}
\title{preprocessInput_snv}
\usage{
preprocessInput_snv(input_data = NULL, ensgene = ensgene,
  reference_genome = ref_genome)
}
\arguments{
\item{input_data}{Mutation data in vcf format.}

\item{ensgene}{Gene table for annotations. A table of Ensembl genes is provided with the package.}

\item{reference_genome}{Reference genome (e.g. BSgenome.Hsapiens.UCSC.hg19)}
}
\description{
Annotating the mutation data with necessary fields for further analysis
}
