% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_preprocess.R
\name{preprocessInput_sv}
\alias{preprocessInput_sv}
\title{preprocessInput_sv}
\usage{
preprocessInput_sv(input_data = NULL, ensgene = ensgene,
  resdir = resdir)
}
\arguments{
\item{input_data}{Table describing somatic structural rearrangements.}

\item{ensgene}{Gene table for annotations. A table of Ensembl genes is provided with the package.}

\item{resdir}{Results directory where graphical outputs should be exported.}
}
\description{
Annotating the mutation data with necessary fields for further analysis
}
