% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_plotting.R
\name{rainfallz_plot}
\alias{rainfallz_plot}
\title{rainfallz_plot}
\usage{
rainfallz_plot(vcf, mutype.col = "mutype", cyto = NULL, resdir = resdir)
}
\arguments{
\item{vcf}{vcf data frame containing the mutations}

\item{mutype.col}{Substitution information column name in vcf}

\item{cyto}{required}

\item{resdir}{Result directory}
}
\description{
Function to plot rainfall plots
}
