% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_plotting.R
\name{RCircos.Get.Start.End.Locations_1}
\alias{RCircos.Get.Start.End.Locations_1}
\title{RCircos.Get.Start.End.Locations_1}
\usage{
RCircos.Get.Start.End.Locations_1(plot.data, plot.width)
}
\arguments{
\item{plot.data}{plot.data}

\item{plot.width}{plot.width}
}
\description{
RCircos.Get.Start.End.Locations_1
}
