% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_plotting.R
\name{RCircos.Heatmap.Plot_1}
\alias{RCircos.Heatmap.Plot_1}
\title{RCircos.Heatmap.Plot_1}
\usage{
RCircos.Heatmap.Plot_1(
  heatmap.data = NULL,
  data.col = NULL,
  track.num = NULL,
  side = c("in", "out"),
  min.value = NULL,
  max.value = NULL,
  inside.pos = NULL,
  outside.pos = NULL,
  genomic.columns = 3,
  is.sorted = TRUE
)
}
\arguments{
\item{heatmap.data}{heatmap.data}

\item{data.col}{data.col}

\item{track.num}{track.num}

\item{side}{side}

\item{min.value}{min.value}

\item{max.value}{max.value}

\item{inside.pos}{inside.pos}

\item{outside.pos}{outside.pos}

\item{genomic.columns}{genomic.columns}

\item{is.sorted}{is.sorted}
}
\description{
RCircos.Heatmap.Plot_1
}
