% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_plotting.R
\name{RCircos.Link.Plot_1}
\alias{RCircos.Link.Plot_1}
\title{RCircos.Link.Plot_1}
\usage{
RCircos.Link.Plot_1(link.data = NULL, track.num = NULL,
  by.chromosome = FALSE, start.pos = NULL, genomic.columns = 3,
  is.sorted = TRUE, lineWidth = rep(1, nrow(link.data)))
}
\arguments{
\item{lineWidth}{}
}
\description{
RCircos.Link.Plot_1
}
