% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_colour_generator_function.R
\name{signature_colour_generator}
\alias{signature_colour_generator}
\title{signature_colour_generator}
\usage{
signature_colour_generator(signature_names = NULL)
}
\arguments{
\item{signature_names}{Character vector of the names of the mutational signatures for which colours are to be generated.}
}
\description{
Generates colours for de novo signatures.
}
\examples{
SBS_colours <- signature_colour_generator(signature_names = rownames(SBS_denovo_sigs))
}
\keyword{Signatures}
