% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_origins_function.R
\name{signature_origins}
\alias{signature_origins}
\title{signature_origins}
\usage{
signature_origins(input = NULL, Type = Type,
  signature_contribution = signatures_exp, input_signatures = NULL)
}
\arguments{
\item{Type}{Mutation type (SBS, DBS, ID or SV)}

\item{signature_contribution}{List of signatures exposure numbers and proportions matrices (output from deconvolution_fit function).}

\item{input_signatures}{Matrix of the input signatures with which the VCF is to be annotated}

\item{vcf}{The input VCF file to which signature origin annotations are to be added.}
}
\description{
Annotates each mutation in a VCF with the signature with which it is most likely associated. E.g. when Type is set to "DBS", each line in the VCF corresponding to a DBS mutation is annotated with the probility that each DBS signature caused it.
}
\examples{
vcf <- signature_origins(vcf=vcf, Type = "SBS", signature_contribution = signatures_exp, input_signatures = SBS_liver_signatures)
}
\keyword{Signatures}
