% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_VCF_function.R
\name{annotate_VCF}
\alias{annotate_VCF}
\title{annotate_VCF}
\usage{
annotate_VCF(vcf = vcf, add_strand_and_SBS_cats = T,
  add_DBS_cats = T, add_ID_cats = T, genome_build = "hg19",
  ref_fasta = NULL, ref_genome = BSgenome.Hsapiens.UCSC.hg19)
}
\arguments{
\item{vcf}{Input VCF to annotate.}

\item{add_strand_and_SBS_cats}{Logical indicating whether or not strand, gene and SBS category annotations are to be added (defaults to TRUE).}

\item{add_DBS_cats}{Logical indicating whether or not DBS category annotations are to be added (defaults to TRUE).}

\item{add_ID_cats}{Logical indicating whether or not Indel category annotations are to be added (defaults to TRUE).}

\item{genome_build}{Enter either "hg19" or "hg38".}

\item{ref_fasta}{File path to FASTA file compatable with input VCF positions and chromosomes.}

\item{ref_genome}{Name of reference genome object. For hg19 data we use the BSgenome.Hsapiens.UCSC.hg19 object, which is loaded into the local environment by library(BSgenome.Hsapiens.UCSC.hg19).}
}
\value{
vcf
}
\description{
Function to add strand, gene and COSMIC mutation category annotations to a VCF. This function can take a long time depending on the number of mutations and how many of the annotation options you have selected. Please be patient!
}
\examples{
vcf <- annotate_VCF(vcf = vcf, ref_genome = BSgenome.Hsapiens.UCSC.hg19, ref_fasta = "~/Documents/Data/Genomes/Homo_sapiens_assembly19.fasta")
}
