% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_plotting_2.R
\name{deconvolution_exposure}
\alias{deconvolution_exposure}
\title{deconvolution_exposure}
\usage{
deconvolution_exposure(
  signature_contribution = signatures_exp,
  rm_samples = NA,
  signature_colours = sig_cols
)
}
\arguments{
\item{signature_contribution}{List of signatures exposure numbers and proportions matrices (output from deconvolution_fit function).}

\item{rm_samples}{List of names of samples to remove (if any). Once the signature contribution of hypermutated samples is known e.g CHC892T in the example dataset), removing these from the grpahical output of this function can make the comparison of the rest of the samples easier.}

\item{signature_colours}{Character vector of R-compatable colours representing each signature to be used graphical outputs. Each signature in input_signatures must have named colour in this vector for grpahical outputs to work.}
}
\description{
Plots the exposure of signatures across the series (bargraphs of the number and proportion of each signature in each sample).
}
\examples{
pdf(file.path(resdir.,"signature_content_plot.pdf"),width=12,height=7)
signature_content_plot <- deconvolution_exposure(signature_contribution = SBS_signatures_exp, rm_samples = "CHC892T", signature_colours = sig_cols)
dev.off()
}
\keyword{Signatures}
