% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_deconvolution.R
\name{deconvolution_fit_SV}
\alias{deconvolution_fit_SV}
\title{deconvolution_fit}
\usage{
deconvolution_fit_SV(vcf = vcf, type = NULL, input_data = data,
  threshold = 5, input_signatures = COSMIC_Signatures,
  sig_cols = mycol, plot = TRUE, resdir = resdir)
}
\arguments{
\item{vcf}{vcf data frame containing the mutations/SVs}

\item{type}{SNV for single nucleotide variant signatures, SV for structural variant signatures}

\item{input_data}{Matrix in mutation type x sample format in proportions}

\item{threshold}{Discard signatures contributing less then this percent of total mutations within each sample}

\item{input_signatures}{Data frame describing the mutational signatures to fit within the provided cohort of samples}

\item{sig_cols}{Character vector indicating the colors representing each signature in graphical outputs. Must match to the total number of provided signatures}

\item{plot}{Logical indicating whether graphical outputs should be generated}

\item{resdir}{Result directory}
}
\description{
Function to identify the weighted combination of the input signatures explaining each tumor's mutational profile
}
