% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dbs_from_vcf_function.R
\name{extract_dbs_from_vcf}
\alias{extract_dbs_from_vcf}
\title{extract_dbs_from_vcf}
\usage{
extract_dbs_from_vcf(vcf = NULL)
}
\arguments{
\item{vcf}{The nput VCF from which DBS mutations are to be extracted}
}
\description{
Extracts lines corresponding to DBS mutations from a VCF. N.B. The VCF must be ordered by sample, CHROM and position for this function to work.
}
\examples{
vcf_dbs <- extract_dbs_from_vcf(vcf = vcf)
}
\keyword{Signatures}
