% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Not_in_function.R
\name{\%!in\%}
\alias{\%!in\%}
\title{"%!in%"}
\usage{
x \%!in\% y
}
\description{
The opposite of "%in%"
}
\examples{
days <- c("monday","tuesday","wednesday","thursday","friday","saturday","sunday")
weekdays <- days[days \%!in\% c("saturday", "sunday")]
}
\keyword{Signatures}
