% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_VCF_functions.R
\name{indel_category_correct}
\alias{indel_category_correct}
\title{indel_category_correct}
\usage{
indel_category_correct(input = NULL)
}
\arguments{
\item{input}{VCF of indel mutations}
}
\description{
Corrects Indel category output from PCAWG7-data-preparation-version-1.5 in the add_ID_cats_ToVCF function. Palimpsest uses the COSMIC indel categories (taken from the Sanger Institute's SigProfiler tool), whereas the PCAWG7-data-preparation-version-1.5 (from the Broad institute) uses slightly different categories.
}
\examples{
vcf_ID <- indel_category_correct(input = vcf)
}
\keyword{Signatures}
