% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_VCF_function.R
\name{order_vcf}
\alias{order_vcf}
\title{order_vcf}
\usage{
order_vcf(vcf, Project_col = NA)
}
\arguments{
\item{vcf}{The VCF to be ordered.}

\item{Project_col}{Name of the Project column in the VCF (if any) (e.g. may contain "ICGC"). If a value is given, the samples are sorted by project first, then sample etc..}
}
\description{
Orders a VCF by Sample name, then genomic position and project (if project argument given).
}
\examples{
vcf <- order_vcf(vcf, Project_col = "Project")
}
\keyword{Signatures}
