% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_VCF_functions.R
\name{palimpsest_addMutationContextToVR}
\alias{palimpsest_addMutationContextToVR}
\title{palimpsest_addMutationContextToVR}
\usage{
palimpsest_addMutationContextToVR(vr = NULL, ref_genome = NULL,
  k = 3, check.strand = FALSE, num_of_DELs = NULL)
}
\value{
vcf
}
\description{
Function used within annotate_VCF to add mutation context to SBS mutations.
}
\examples{
vr3 <- palimpsest_addMutationContextToVR(vr = vr, ref_genome = ref_genome, k = 3, num_of_DELs = delet_this)
}
