% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV_functions.R
\name{preprocessInput_sv}
\alias{preprocessInput_sv}
\title{preprocessInput_sv}
\usage{
preprocessInput_sv(
  input_data = NULL,
  Sample.col = "Sample",
  CHROM_1.col = "CHROM_1",
  CHROM_2.col = "CHROM_2",
  POS_1.col = "POS_1",
  POS_2.col = "POS_2",
  type.col = "Type",
  resdir = resdir,
  genome_build = "hg19"
)
}
\arguments{
\item{input_data}{Table describing somatic structural rearrangements.}

\item{resdir}{Results directory where graphical outputs should be exported.}

\item{ensgene}{Gene table for annotations. A table of Ensembl genes is provided with the package.}
}
\description{
Annotating the mutation data with necessary fields for further analysis
}
