% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palimpsest_plotting.R
\name{RCircos.Heatmap.Plot_1}
\alias{RCircos.Heatmap.Plot_1}
\title{RCircos.Heatmap.Plot_1}
\usage{
RCircos.Heatmap.Plot_1(heatmap.data = NULL, data.col = NULL,
  track.num = NULL, side = c("in", "out"), min.value = NULL,
  max.value = NULL, inside.pos = NULL, outside.pos = NULL,
  genomic.columns = 3, is.sorted = TRUE)
}
\arguments{
\item{is.sorted}{}
}
\description{
RCircos.Heatmap.Plot_1
}
